+++ cloud-siem-20241212/alibabacloud_cloud_siem20241212/models/_list_data_set_records_request.py
@@ -8,9 +8,12 @@ class ListDataSetRecordsRequest(DaraModel):
     def __init__(
         self,
         data_set_id: str = None,
+        filter: str = None,
         lang: str = None,
         max_results: int = None,
         next_token: str = None,
+        order: str = None,
+        order_field: str = None,
         page_number: int = None,
         page_size: int = None,
         region_id: str = None,
@@ -18,9 +21,12 @@ class ListDataSetRecordsRequest(DaraModel):
     ):
         # This parameter is required.
         self.data_set_id = data_set_id
+        self.filter = filter
         self.lang = lang
         self.max_results = max_results
         self.next_token = next_token
+        self.order = order
+        self.order_field = order_field
         # This parameter is required.
         self.page_number = page_number
         # This parameter is required.
@@ -39,6 +45,9 @@ class ListDataSetRecordsRequest(DaraModel):
         if self.data_set_id is not None:
             result['DataSetId'] = self.data_set_id
 
+        if self.filter is not None:
+            result['Filter'] = self.filter
+
         if self.lang is not None:
             result['Lang'] = self.lang
 
@@ -47,6 +56,12 @@ class ListDataSetRecordsRequest(DaraModel):
 
         if self.next_token is not None:
             result['NextToken'] = self.next_token
+
+        if self.order is not None:
+            result['Order'] = self.order
+
+        if self.order_field is not None:
+            result['OrderField'] = self.order_field
 
         if self.page_number is not None:
             result['PageNumber'] = self.page_number
@@ -67,6 +82,9 @@ class ListDataSetRecordsRequest(DaraModel):
         if m.get('DataSetId') is not None:
             self.data_set_id = m.get('DataSetId')
 
+        if m.get('Filter') is not None:
+            self.filter = m.get('Filter')
+
         if m.get('Lang') is not None:
             self.lang = m.get('Lang')
 
@@ -75,6 +93,12 @@ class ListDataSetRecordsRequest(DaraModel):
 
         if m.get('NextToken') is not None:
             self.next_token = m.get('NextToken')
+
+        if m.get('Order') is not None:
+            self.order = m.get('Order')
+
+        if m.get('OrderField') is not None:
+            self.order_field = m.get('OrderField')
 
         if m.get('PageNumber') is not None:
             self.page_number = m.get('PageNumber')