+++ cloud-siem-20241212/alibabacloud_cloud_siem20241212/models/_list_detection_rules_response_body.py
@@ -94,8 +94,10 @@ class ListDetectionRulesResponseBodyDetectionRules(DaraModel):
     def __init__(
         self,
         alert_att_ck: str = None,
+        alert_att_ck_mapping: str = None,
         alert_description: str = None,
         alert_level: str = None,
+        alert_level_mapping: str = None,
         alert_name: str = None,
         alert_schema_id: str = None,
         alert_tactic_id: str = None,
@@ -103,6 +105,7 @@ class ListDetectionRulesResponseBodyDetectionRules(DaraModel):
         alert_threshold_group: str = None,
         alert_threshold_period: str = None,
         alert_type: str = None,
+        alert_type_mapping: str = None,
         create_time: int = None,
         detection_expression_content: str = None,
         detection_expression_type: str = None,
@@ -128,8 +131,10 @@ class ListDetectionRulesResponseBodyDetectionRules(DaraModel):
         update_time: int = None,
     ):
         self.alert_att_ck = alert_att_ck
+        self.alert_att_ck_mapping = alert_att_ck_mapping
         self.alert_description = alert_description
         self.alert_level = alert_level
+        self.alert_level_mapping = alert_level_mapping
         self.alert_name = alert_name
         self.alert_schema_id = alert_schema_id
         self.alert_tactic_id = alert_tactic_id
@@ -137,6 +142,7 @@ class ListDetectionRulesResponseBodyDetectionRules(DaraModel):
         self.alert_threshold_group = alert_threshold_group
         self.alert_threshold_period = alert_threshold_period
         self.alert_type = alert_type
+        self.alert_type_mapping = alert_type_mapping
         self.create_time = create_time
         self.detection_expression_content = detection_expression_content
         self.detection_expression_type = detection_expression_type
@@ -177,11 +183,17 @@ class ListDetectionRulesResponseBodyDetectionRules(DaraModel):
         if self.alert_att_ck is not None:
             result['AlertAttCk'] = self.alert_att_ck
 
+        if self.alert_att_ck_mapping is not None:
+            result['AlertAttCkMapping'] = self.alert_att_ck_mapping
+
         if self.alert_description is not None:
             result['AlertDescription'] = self.alert_description
 
         if self.alert_level is not None:
             result['AlertLevel'] = self.alert_level
+
+        if self.alert_level_mapping is not None:
+            result['AlertLevelMapping'] = self.alert_level_mapping
 
         if self.alert_name is not None:
             result['AlertName'] = self.alert_name
@@ -203,6 +215,9 @@ class ListDetectionRulesResponseBodyDetectionRules(DaraModel):
 
         if self.alert_type is not None:
             result['AlertType'] = self.alert_type
+
+        if self.alert_type_mapping is not None:
+            result['AlertTypeMapping'] = self.alert_type_mapping
 
         if self.create_time is not None:
             result['CreateTime'] = self.create_time
@@ -282,11 +297,17 @@ class ListDetectionRulesResponseBodyDetectionRules(DaraModel):
         if m.get('AlertAttCk') is not None:
             self.alert_att_ck = m.get('AlertAttCk')
 
+        if m.get('AlertAttCkMapping') is not None:
+            self.alert_att_ck_mapping = m.get('AlertAttCkMapping')
+
         if m.get('AlertDescription') is not None:
             self.alert_description = m.get('AlertDescription')
 
         if m.get('AlertLevel') is not None:
             self.alert_level = m.get('AlertLevel')
+
+        if m.get('AlertLevelMapping') is not None:
+            self.alert_level_mapping = m.get('AlertLevelMapping')
 
         if m.get('AlertName') is not None:
             self.alert_name = m.get('AlertName')
@@ -308,6 +329,9 @@ class ListDetectionRulesResponseBodyDetectionRules(DaraModel):
 
         if m.get('AlertType') is not None:
             self.alert_type = m.get('AlertType')
+
+        if m.get('AlertTypeMapping') is not None:
+            self.alert_type_mapping = m.get('AlertTypeMapping')
 
         if m.get('CreateTime') is not None:
             self.create_time = m.get('CreateTime')